4.21.- expansión de la lista de blogs, paso 9
El blog solo puede ser eliminado por el usuario que agrego el blog

+++ controllers/blogs.js
@@ -60,9 +60,20 @@ blogsRouter.post('/', async (request, response) => {
 
 //Delete
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = await request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({
+      error: 'no permission'
+    })
+  }
+
 })
+
 
 //Update
 blogsRouter.put('/:id', async (request, response) => {
